requirements.txt
streamlit>=1.31.0
pandas>=2.1.4
openpyxl>=3.1.2
plotly>=5.18.0
requests>=2.31.0
Pillow>=10.2.0
pyarrow>=15.0.0